backend/requirements.txt
# FastAPI Framework
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
greenlet==3.0.1

# Cache and Task Queue
redis==5.0.1
celery==5.3.6

# HTTP Client
requests==2.31.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
PyJWT==2.8.0
pytz==2023.3.post1
emoji==2.10.1
cryptography==42.0.2
jdatetime==4.1.1
persiantools==3.0.1

# Payment Integration
zarinpal==1.5.0

# Monitoring & Logging
sentry-sdk==1.32.0
prometheus-client==0.17.1
python-json-logger==2.0.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Additional Dependencies
jinja2==3.1.3
pillow==10.2.0
python-i18n==0.3.9
rich==13.7.0 